% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saceestimwvarglm.R
\name{saceglm}
\alias{saceglm}
\title{Estimation of SACE in Cluster Randomized Trials Using Marginal Model for Survival Event}
\usage{
saceglm(
  data,
  trt = "A",
  surv = "S",
  out = "Y",
  clustid = "Id",
  indv = "X",
  crobust = T,
  set1 = T,
  set2 = T,
  conf = 0.95,
  boot = F,
  dfc = F,
  varpen = F,
  iters = 200
)
}
\arguments{
\item{data}{Data frame containing all data required for estimation with user-specified names below.}

\item{trt}{A named \code{character} specifying treatment variable. Default is "A".}

\item{surv}{A named \code{character} specifying survival status, where survival through study is indicated by 1 and death by 0. Default is "S".}

\item{out}{A named \code{character} specifying non-mortal outcome. Default is "Y".}

\item{clustid}{A named \code{character} specifying cluster membership. Default is "Id".}

\item{indv}{A named \code{character} vector for covariates to be treated as fixed effects. Group-level variables can be included but they must be defined
for each individual in the group. Default is "X".}

\item{crobust}{A \code{logical} specifying whether cluster-robust variance estimate is provided (1,1 submatrix equal to \verb{vcov(geepack::geeglm(formula,family="binomial))}).
If set to \code{F}, the clustering structure is completely ignored in the sandwich variance expression (1,1 submatrix equal to \code{sandwich(glm(formula,family="binomial"))}). Default is \code{T}.}

\item{set1}{A \code{logical} argument for whether identified estimator uses Set 1 Assumptions. Default is \code{T}.}

\item{set2}{A \code{logical} argument for whether identified estimator uses Set 2 Assumptions. Default is \code{F}.
If \code{set1=T} and \code{set2=T}, function will provide results for both estimators.}

\item{conf}{A \code{numeric} argument in the interval (0,1) for \% confidence interval. Default is \code{.95}.}

\item{boot}{A \code{logical} argument for variance estimation. If \code{boot=F}, variance is estimated}

\item{dfc}{A \code{logical} argument for a degrees of freedom adjustment to variance if \code{boot=F}. Default is \code{F}.}

\item{varpen}{A \code{logical} argument for an extra degree of freedom penalty for accounting for estimating variance of random intercept if \code{boot=F}. Default is \code{F} and should not be set to \code{T}.}

\item{iters}{A \code{double} for number of bootstrap samples to be taken when \code{boot=T}. Default is \code{iters=200}. This argument is ignored when \code{boot=F}.}
}
\value{
A named \code{double} including point estimates, estimates of variance, and confidence intervals.
}
\description{
This function provides interval estimates of the causal estimand SACE for cluster randomized
trials based (CRTs) for the SSW and PSW estimators derived from identification assumptions
as described in Isenberg et al.\if{html}{\out{<sup>}}1\if{html}{\out{</sup>}}, where all survival status models are fit
using logistic regression for binary data. User has the option
to choose how variances are estimated, and corresponding confidence intervals are constructed.
By default, this function provides the variance of the asymptotic distribution of
these estimators using the theory of M-estimation\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}}, resulting in a cluster-robust variance estimate\if{html}{\out{<sup>}}3\if{html}{\out{</sup>}}.
Alternatively, the user may select to estimate variance using the non-parametric cluster bootstrap option\if{html}{\out{<sup>}}4\if{html}{\out{</sup>}}.
}
\references{
\enumerate{
\item{Isenberg, D., Harhay, M.O., Mitra, N. and Li, F., 2025. Weighting methods for truncation by death in cluster-randomized trials. \emph{Statistical Methods in Medical Research, 34(3)}, pp.473-489.}
\item{Stefanski L.A. and Boos D.D., 2002. The calculus of M-estimation. \emph{The American Statistician, 56(1)}, pp.29–38.}
\item{Liang, K.Y. and Zeger, S.L., 1986. Longitudinal data analysis using generalized linear models. \emph{Biometrika, 73(1)}, pp.13-22.}
\item{Field C.A. and Welsh A.H., 2007. Bootstrapping clustered data. \emph{Journal of the Royal Statistical Society Series B: Statistical Methodology, 69(3)}, pp.369–390.}
}
}
